+++ src/gen-xml.ts
@@ -438,7 +438,24 @@ function slideObjectToXml(slide: ISlideLib | ISlideLayout): string {
 				strSlideXml += '<p:sp>'
 
 				// B: The addition of the "txBox" attribute is the sole determiner of if an object is a shape or textbox
-				strSlideXml += `<p:nvSpPr><p:cNvPr id="${idx + 2}" name="Object${idx + 1}"/>`
+				strSlideXml += `<p:nvSpPr><p:cNvPr id="${idx + 2}" name="Object${idx + 1}">`
+				// <Hyperlink>
+				if (slideItemObj.options.hyperlink && slideItemObj.options.hyperlink.url)
+					strSlideXml +=
+						'<a:hlinkClick r:id="rId' +
+						slideItemObj.options.hyperlink.rId +
+						'" tooltip="' +
+						(slideItemObj.options.hyperlink.tooltip ? encodeXmlEntities(slideItemObj.options.hyperlink.tooltip) : '') +
+						'"/>'
+				if (slideItemObj.options.hyperlink && slideItemObj.options.hyperlink.slide)
+					strSlideXml +=
+						'<a:hlinkClick r:id="rId' +
+						slideItemObj.options.hyperlink.rId +
+						'" tooltip="' +
+						(slideItemObj.options.hyperlink.tooltip ? encodeXmlEntities(slideItemObj.options.hyperlink.tooltip) : '') +
+						'" action="ppaction://hlinksldjump"/>'
+				// </Hyperlink>
+				strSlideXml += '</p:cNvPr>'
 				strSlideXml += '<p:cNvSpPr' + (slideItemObj.options && slideItemObj.options.isTextBox ? ' txBox="1"/>' : '/>')
 				strSlideXml += `<p:nvPr>${slideItemObj.type === 'placeholder' ? genXmlPlaceholder(slideItemObj) : genXmlPlaceholder(placeholderObj)}</p:nvPr>`
 				strSlideXml += '</p:nvSpPr><p:spPr>'